+++ ACP3/Modules/ACP3/Guestbook/Validation/FormValidation.php
@@ -2,7 +2,6 @@
 namespace ACP3\Modules\ACP3\Guestbook\Validation;
 
 use ACP3\Core;
-use ACP3\Modules\ACP3\Captcha\Validation\ValidationRules\CaptchaValidationRule;
 use ACP3\Modules\ACP3\Guestbook\Validation\ValidationRules\FloodBarrierValidationRule;
 use ACP3\Modules\ACP3\Newsletter;
 
@@ -71,13 +70,6 @@ class FormValidation extends Core\Validation\AbstractFormValidation
                     'data' => $formData,
                     'field' => 'message',
                     'message' => $this->translator->t('system', 'message_to_short')
-                ])
-            ->addConstraint(
-                CaptchaValidationRule::class,
-                [
-                    'data' => $formData,
-                    'field' => 'captcha',
-                    'message' => $this->translator->t('captcha', 'invalid_captcha_entered')
                 ]);
 
         if (!empty($formData['mail'])) {
@@ -110,6 +102,8 @@ class FormValidation extends Core\Validation\AbstractFormValidation
                     ]
                 );
         }
+
+        $this->validator->dispatchValidationEvent('captcha.validation.validate_captcha', $formData);
 
         $this->validator->validate();
     }